+++ stepmania/src/arch/MovieTexture/MovieTexture_FFMpeg.cpp
@@ -412,7 +412,7 @@ int MovieDecoder_FFMpeg::DecodePacket( float fTargetTime )
 		{
 			if (m_Packet.dts != int64_t(AV_NOPTS_VALUE))
 			{
-				m_fPTS = m_Packet.dts * av_q2d(m_pStream->time_base);
+				m_fPTS = float( m_Packet.dts * av_q2d(m_pStream->time_base) );
 
 				/* dts is the timestamp of the first frame in this packet. Only use it once;
 				 * if we get more than one frame from the same packet (eg. f;lushing the last
